pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cloud-monitor-agent"
version = "2.0.0"
description = "Agente ADK para monitoramento de recursos GCP via ferramentas Python nativas"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-monitoring>=2.22.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cloud_logging_agent*", "api*"]